failing swift tests
