+++ crates/core/src/live_socket/tests/error.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::error::*;
 
 #[tokio::test]
 async fn dead_render_error() {
